# openapi.yml
openapi: "3.1.3"
info:
  title: Developer Loop API
  description: API to manage and execute shell scripts with metadata, execution, and history tracking.
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /:
    get:
      summary: Serve front page UI
      responses:
        '200':
          description: Front page

  /public:
    get:
      summary: Serve static assets
      responses:
        '200':
          description: Static assets

  /api/actions/scripts/load:
    post:
      summary: Load or refresh scripts from folder(s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folders:
                  type: array
                  items:
                    type: string
                  example: ["/scripts/path"]
      responses:
        '200':
          description: Scripts loaded successfully

  /api/actions/exec/scripts/{id}:
    post:
      summary: Execute a script by ID, stream stdout/stderr, and save to history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: incognito
          in: query
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                args:
                  type: array
                  items:
                    type: string
                  example: ["--version"]
                env:
                  type: object
                  additionalProperties:
                    type: string
                  example: { COUNT: "2" }
                command:
                  type: string
                  example: "bash"
      responses:
        '200':
          description: Execution started

  /api/scripts:
    get:
      summary: Get all scripts metadata
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of scripts metadata

  /api/scripts/{id}:
    get:
      summary: Get metadata of a script by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Script metadata

    delete:
      summary: Delete script metadata and optionally the file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Script deleted

    patch:
      summary: Open script in editor (default or custom)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Script opened in editor

  /api/history/scripts/{id}:
    get:
      summary: Get execution history for a script (excluding output)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of script history

  /api/history/{id}:
    get:
      summary: Get history by ID (include output)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: History details

    delete:
      summary: Delete specific history entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: History deleted